requirements.txt
yfinance==0.1.55
numpy==1.20.1
streamlit==0.80.0
matplotlib==3.3.4
pandas==1.2.3
Pillow==8.2.0
